+++ octomap/src/Pointcloud.cpp
@@ -284,7 +284,7 @@ namespace octomap {
 
   std::istream& Pointcloud::readBinary(std::istream &s) {
 
-    unsigned int pc_size = 0;
+    size_t pc_size = 0;
     s.read((char*)&pc_size, sizeof(pc_size));
     OCTOMAP_DEBUG("Reading %d points from binary file...", pc_size);
